src: tags.yaml: allow time ranges with empty STEP
Allow empty STEPs in a time range to denote a timestamp series
consisting of only the START and END timestamps.  This is enables
fetching peak demands over a single period of time. In all other cases
so far, START::END would be equivalent to END,START but the latter
would not yield a peak demand wheras the former would.