bin/build_openapi.py: add support for redocly "Reunite" link format
Link anchors are quite different for the new "Reunit" redoc product so
this commit includes updates to all operation (op:PATH:METHOD) links.
Fortunately, there aren't too many of these and the new format allows
us to generate both the old and new formats.

For this to work, a couple of operationId tags had to be updated as
well so they fit with the new requirements.

The final openapi.yaml is now placed in the "output" directory
(instead of the top-level directory).  This makes it easier to import
just the output into Reunite.

 Conflicts:
	bin/build_openapi.py
	output/openapi.yaml